+++ src/build_context.rs
@@ -83,6 +83,8 @@ pub struct BuildContext {
     pub bridge: BridgeModel,
     /// Whether this project is pure rust or rust mixed with python
     pub project_layout: ProjectLayout,
+    /// The path to pyproject.toml. Required for the source distribution
+    pub pyproject_toml_path: PathBuf,
     /// Parsed pyproject.toml if any
     pub pyproject_toml: Option<PyProjectToml>,
     /// Python Package Metadata 2.1
@@ -176,19 +178,10 @@ impl BuildContext {
         fs::create_dir_all(&self.out)
             .context("Failed to create the target directory for the source distribution")?;
 
-        let include_cargo_lock = self.cargo_options.locked || self.cargo_options.frozen;
         match self.pyproject_toml.as_ref() {
             Some(pyproject) => {
-                let sdist_path = source_distribution(
-                    &self.out,
-                    &self.metadata21,
-                    &self.manifest_path,
-                    &self.cargo_metadata,
-                    pyproject.sdist_include(),
-                    include_cargo_lock,
-                    self.project_layout.data.as_deref(),
-                )
-                .context("Failed to build source distribution")?;
+                let sdist_path = source_distribution(self, pyproject)
+                    .context("Failed to build source distribution")?;
                 Ok(Some((sdist_path, "source".to_string())))
             }
             None => Ok(None),